Refactor CLI options out into separate file.
